// proyecto-antiguo/frontend/src/types/api.ts
export interface SearchResult {
  id_producto?: number | null;
  producto: string;
  pvu?: number | null;
  pcu?: number | null;
  unidades?: number | null;
  licitacion_nombre?: string | null;
  numero_expediente?: string | null;
  proveedor?: string | null;
}